The objective of this code is to perform voltage regulation using Genetic Algorithm (GA) by obtaining optimal tap number for transformers and optimal switching status of the units of capacitor banks (CBs). The optimization is repeated after a chosen time interval for the considered total simulation time.
GA Outputs:
x: Optimal value for the uknown variables. It is a row vector of order (1 x nVar).
fval: Optimal value of the fitness function.
exitflag: An integer code denoting the reason for termination of the iterations.
output: It is a structure providing information about the optimization process. It has predefined fields as shown below:
        a) probelmtype - Problem type, i.e., unconstrained, boundconstraints, linearconstraints, nonlinearconstr, integerconstraints.
        b) rngstate - State of the MATLAB random number generator, just before the algorithm started.
        c) generations - Number of generations computed.
        d) funccount - Number of evaluations of the fitness function.
        e) message — Reason the algorithm terminated.
        f) maxconstraint — Maximum constraint violation, if any.
        g) hybridflag — Exit flag from the hybrid function.
population: It is a matrix of order (popsize x nVar) where popsize is the population size and nVar is the number of unknown variables.
scores: Value of the fitness function for each individual. It is a column vector of size (popsize x 1).